+++ CHANGES.md
@@ -1,6 +1,13 @@
 History
 =======
 
+### Version 0.0.7 (next version)
+
+* Catch up puppeteer v0.10.2
+    * Add `Page.querySelectorEval` (`Page.$eval` in puppeteer)
+    * Deprecate `ElementHandle.attribute`
+    * Add `Touchscreen` class and implement `Page.tap` and `ElementHandle.tap`
+
 ### Version 0.0.6 (2017-09-02)
 
 * Accept keyword arguments for options